Move Discord IDs and command checks to separate files
