+++ csmgr/discord_ids.py
@@ -1,0 +1,7 @@
+COG_CREATOR_ROLE_ID = 529359322140901377
+COG_SUPPORT_SERVER_ID = 240154543684321280
+CORE_DEV_ROLE_ID = 595432395214422017
+OTHERCOGS_ID = 240212783503900673
+QA_ROLE_ID = 477939350399746105
+SENIOR_COG_CREATOR_ROLE_ID = 533275636387807262
+V3_COG_SUPPORT_CATEGORY_ID = 533567794978226176
